backend/requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.8
python-dotenv==1.0.0
pydantic==1.10.5
python-jose
passlib
alembic
python-multipart